+++ example/interrupts.dart
@@ -27,18 +27,18 @@ main() async {
     print('${stopwatch.elapsedMilliseconds} value: ${event.value}');
 
     // Toggle the LED
-    ledPin.value = event.value == 1 ? 0 : 1;
+    ledPin.value = !event.value;
   });
 
   // Turn on the LED
-  ledPin.value = 1;
+  ledPin.value = true;
 
   // Cancel listening for interrupts after 1/4 second
   new Timer(new Duration(milliseconds: 250), () {
     subscription.cancel();
 
     // Ensure LED is off
-    ledPin.value = 0;
+    ledPin.value = false;
   });
 }