+++ test/all.dart
@@ -22,8 +22,8 @@ runTests() {
 
   test('pins used', () {
     // Ensure LEDs are off
-    pin(1, Mode.output)..value = 0;
-    pin(3, Mode.output)..value = 0;
+    pin(1, Mode.output)..value = false;
+    pin(3, Mode.output)..value = false;
     // Print pin state changes
     if (recording != null) recording.printUsage();
   });